pyproject.toml
[tool.poetry]
name = "nbltg"
version = "0.1.0"
description = ""
authors = ["Rayman <ggammi1007@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "2.25.2"
vk-api = "^11.9.9"
python-dotenv = "^1.0.1"
aiosqlite = "^0.19.0"
bs4 = "^0.0.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
